Port some assembly in ID_CA.C. Fix a bunch of assembly operand size mismatches by changing ints to shorts. Fix invalid 16-bit assembly indirect addressing by using E versions of registers.

+++ WL_GAME.C
@@ -936,14 +936,14 @@ char	demoname[13] = "DEMO?.";
 
 void FinishDemoRecord (void)
 {
-	long	length,level;
+	int	length,level;
 
 	demorecord = False;
 
 	length = demoptr - (char *)demobuffer;
 
 	demoptr = ((char *)demobuffer)+1;
-	*(unsigned *)demoptr = length;
+	*(unsigned int *)demoptr = length;
 
 	CenterWindow(24,3);
 	PrintY+=6;
